refactor(175): update architecture to put test file for useToaster
