+++ src/common/composables/index.ts
@@ -1,2 +1,3 @@
 export * from './use-language-switcher'
 export * from './use-navigation'
+export * from './use-toaster'
